pyproject.toml
[tool.poetry]
name = "trip-planner-yt"
version = "0.1.0"
description = "Simple ai trip planner"
authors = ["Joao Moura, Brandon Hancock"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.11.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"
openai = "1.7.1"
langchain-community = "0.0.10"
langchain-core = "0.1.8"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"